add url params for analyzer y-axes
