+++ projects/tools/src/lib/analyzer.service.ts
@@ -18,7 +18,9 @@ export class AnalyzerService {
     analyzerTableDates: [],
     analyzerSeries: [],
     yAxes: [],
-    highstockSeriesOptions: []
+    highstockSeriesOptions: [],
+    y0Series: null,
+    y1Series: null
   };
   public embedData = {
     analyzerTableDates: [],
@@ -84,7 +86,9 @@ export class AnalyzerService {
 
   getAnalyzerData(aSeries, noCache: boolean, y0Series: string, y1Series: string) {
     this.analyzerData.analyzerSeries = [];
+    console.log('ASERIES', aSeries)
     const ids = aSeries.map(s => s.id).join();
+    console.log('ids', ids)
     this.apiService.fetchPackageAnalyzer(ids, noCache).subscribe((results) => {
       const series = results.series;
       series.forEach((s) => {
@@ -94,14 +98,17 @@ export class AnalyzerService {
       this.createAnalyzerTable(this.analyzerData.analyzerSeries);
       this.checkAnalyzerChartSeries();
       this.analyzerData.yAxes = this.setYAxes(this.analyzerData.analyzerSeries, y0Series, y1Series);
-      this.analyzerData.highstockSeriesOptions = this.formatSeriesData(this.analyzerData.analyzerSeries, this.analyzerData.analyzerTableDates, this.analyzerData.yAxes);
+      console.log('Y0SERIES', y0Series)
+      this.analyzerData.y0Series = y0Series ? y0Series.split('-').map(s => +s) : null;
+      this.analyzerData.y1Series = y1Series ? y1Series.split('-').map(s => +s) : null;
       console.log('ANALYZER DATA', this.analyzerData)
     });
     return observableForkJoin([observableOf(this.analyzerData)]);
   }
 
   formatSeriesForAnalyzer = (series, aSeries) => {
-    let decimals;
+    console.log('FORMAT SERIES', series);
+    console.log('FORMAT aSERIES', aSeries)
     const aSeriesMatch = aSeries.find(a => a.id === series.id);
     const seriesData = {
       seriesDetail: series,
@@ -134,7 +141,6 @@ export class AnalyzerService {
     seriesData.displayName = this.formatDisplayName(abbreviatedNameDetails);
     seriesData.chartDisplayName = this.formatDisplayName(chartNameDetails);
     seriesData.saParam = series.seasonalAdjustment !== 'not_seasonally_adjusted';
-    decimals = series.decimals ? series.decimals : 1;
     seriesData.currentGeo = series.geography;
     seriesData.currentFreq = { freq: series.frequencyShort, label: series.frequency };
     seriesData.observations = series.seriesObservations;
@@ -161,67 +167,6 @@ export class AnalyzerService {
     }
     return seriesData;
   }
-  /* formatSeriesForAnalyzer = (series, showInChart: boolean) => {
-    const aSeriesMatch = aSeries.find(a => a.id === series.id);
-    const seriesData = {
-      seriesDetail: series,
-      currentGeo: {} as Geography,
-      currentFreq: {} as Frequency,
-      chartData: {},
-      displayName: '',
-      chartDisplayName: '',
-      seriesTableData: [],
-      error: null,
-      saParam: false,
-      noData: '',
-      observations: { transformationResults: [], observationStart: '', observationEnd: '' },
-      showInChart: aSeriesMatch.showInChart,
-      title: ''
-    };
-    const abbreviatedNameDetails = {
-      title: series.title,
-      geography: series.geography.shortName,
-      frequency: series.frequency,
-      seasonalAdjustment: series.seasonalAdjustment,
-      units: series.unitsLabelShort ? series.unitsLabelShort : series.unitsLabel
-    };
-    const chartNameDetails = {
-      title: series.title,
-      geography: series.geography.shortName,
-      frequency: series.frequency,
-      seasonalAdjustment: series.seasonalAdjustment,
-      units: series.unitsLabelShort ? series.unitsLabelShort : series.unitsLabel
-    };
-    seriesData.displayName = this.formatDisplayName(abbreviatedNameDetails);
-    seriesData.chartDisplayName = this.formatDisplayName(chartNameDetails);
-    seriesData.title = series.title,
-    seriesData.saParam = series.seasonalAdjustment !== 'not_seasonally_adjusted';
-    seriesData.currentGeo = series.geography;
-    seriesData.currentFreq = { freq: series.frequencyShort, label: series.frequency };
-    seriesData.observations = series.seriesObservations;
-    const levelDates = seriesData.observations.transformationResults[0].dates;
-    const obsStart = seriesData.observations.observationStart;
-    const obsEnd = seriesData.observations.observationEnd;
-    const dateArray = [];
-    if (levelDates) {
-      const pseudoZones = [];
-      const level = seriesData.observations.transformationResults[0].values;
-      if (level.pseudoHistory) {
-        level.pseudoHistory.forEach((obs, index) => {
-          if (obs && !level.pseudoHistory[index + 1]) {
-            pseudoZones.push({ value: Date.parse(level.dates[index]), dashStyle: 'dash', color: '#7CB5EC', className: 'pseudoHistory' });
-          }
-        });
-      }
-      // Use to format dates for table
-      this.helperService.createDateArray(obsStart, obsEnd, seriesData.currentFreq.freq, dateArray);
-      const levelChartData = this.createSeriesChartData(seriesData.observations.transformationResults[0], dateArray);
-      seriesData.chartData = { level: levelChartData, dates: dateArray, pseudoZones };
-    } else {
-      seriesData.noData = 'Data not available';
-    }
-    return seriesData;
-  } */
 
   checkFrequencies = (series) => {
     const freqs = series.map((s) => s.currentFreq.freq);
@@ -325,27 +270,18 @@ export class AnalyzerService {
       this.analyzerSeries.splice(seriesExist, 1);
       this.analyzerData.analyzerSeries.splice(this.analyzerData.analyzerSeries.findIndex(series => series.seriesDetail.id === seriesId), 1);
       this.analyzerData.analyzerTableDates = this.createAnalyzerTableDates(this.analyzerData.analyzerSeries);
+      this.analyzerData.yAxes = this.setYAxes(this.analyzerData.analyzerSeries, this.analyzerData.y0Series, this.analyzerData.y1Series);
       this.analyzerSeriesCount.next(this.analyzerSeries.length);
     }
     if (seriesExist < 0) {
       this.analyzerSeries.push({ id: seriesId });
       this.analyzerSeriesCount.next(this.analyzerSeries.length);
       this.analyzerData.analyzerTableDates = this.createAnalyzerTableDates(this.analyzerData.analyzerSeries);
+      this.analyzerData.yAxes = this.setYAxes(this.analyzerData.analyzerSeries, this.analyzerData.y0Series, this.analyzerData.y1Series);
     }
   }
 
   setYAxes = (series, y0Series, y1Series) => {
-    let y0;
-    let y1;
-    if (y0Series) {
-      y0 = y0Series.split('-').map(s => +s);
-    }
-    if (y1Series) {
-      y1 = y1Series.split('-').map(s => +s);
-    }
-    console.log('SET Y AXES y0', y0);
-    console.log('SET Y AXES y1', y1)
-
     // Group series by their units
     // i.e., If series with 2 different units have been selected, draw a y-axis for each unit
     const axisIds = {
@@ -353,17 +289,15 @@ export class AnalyzerService {
       yAxis1: []
     };
     series.reduce((obj, serie) => {
-      console.log('SERIES REDUCE', serie);
-      if (y0 && y0.includes(serie.seriesDetail.id)) {
+      if (y0Series && y0Series.includes(serie.seriesDetail.id)) {
         obj.yAxis0.push(serie);
         return obj;
       }
-      if (y1 && y1.includes(serie.seriesDetail.id)) {
-        console.log('Y1 TRUE')
+      if (y1Series && y1Series.includes(serie.seriesDetail.id)) {
         obj.yAxis1.push(serie);
         return obj;
       }
-      if (!y0 || !y1) {
+      if (!y0Series || !y1Series) {
         if (!obj.yAxis0.length) {
           obj.yAxis0.push(serie);
           return obj;
@@ -379,9 +313,7 @@ export class AnalyzerService {
       }
     }, axisIds);
     const yAxes = Object.keys(axisIds).map((axis, index) => {
-      const atLeastOneSeriesVisible = axisIds[axis].find(s => s.showInChart);
-      console.log('atLeastOneVisible', atLeastOneSeriesVisible);
-      console.log('axis', axisIds[axis])
+      const visibleSeries = axisIds[axis].find(s => s.showInChart);
       return {
         labels: {
           formatter() {
@@ -390,7 +322,7 @@ export class AnalyzerService {
         },
         id: axis,
         title: {
-          text: atLeastOneSeriesVisible ? atLeastOneSeriesVisible.seriesDetail.unitsLabelShort : null
+          text: visibleSeries ? visibleSeries.seriesDetail.unitsLabelShort : null
         },
         opposite: index === 0 ? false : true,
         minPadding: 0,
@@ -398,68 +330,10 @@ export class AnalyzerService {
         minTickInterval: 0.01,
         showEmpty: false,
         series: axisIds[axis],
-        visible: atLeastOneSeriesVisible ? true : false
+        visible: visibleSeries ? true : false
       };
     });
     console.log('yAxes', yAxes)
     return yAxes;
-  }
-
-  formatSeriesData = (series: Array<any>, dates: Array<any>, yAxes: Array<any>) => {
-    console.log('FORMAT SERIES DATA', yAxes);
-    const chartSeries = series.map((serie) => {
-      const axis = yAxes ? yAxes.find(y => y.series.some(s => s.seriesDetail.id === serie.seriesDetail.id)) : null;
-      return {
-        className: serie.seriesDetail.id,
-        name: serie.chartDisplayName,
-        tooltipName: serie.seriesDetail.title,
-        data: serie.chartData.level,
-        yAxis: axis ? axis.id : null,
-        decimals: serie.seriesDetail.decimals,
-        frequency: serie.seriesDetail.frequencyShort,
-        geography: serie.seriesDetail.geography.name,
-        includeInDataExport: serie.showInChart ? true : false,
-        showInLegend: serie.showInChart ? true : false,
-        showInNavigator: false,
-        events: {
-          legendItemClick() {
-            return false;
-          }
-        },
-        unitsLabelShort: serie.seriesDetail.unitsLabelShort,
-        seasonallyAdjusted: serie.seriesDetail.seasonalAdjustment === 'seasonally_adjusted',
-        dataGrouping: {
-          enabled: false
-        },
-        pseudoZones: serie.chartData.pseudoZones,
-        visible: serie.showInChart ? true : false
-      };
-    });
-    chartSeries.push({
-      className: 'navigator',
-      data: dates.map(d => [Date.parse(d.date), null]),
-      decimals: null,
-      tooltipName: '',
-      frequency: null,
-      geography: null,
-      yAxis: 'yAxis0',
-      dataGrouping: {
-        enabled: false
-      },
-      showInLegend: false,
-      showInNavigator: true,
-      includeInDataExport: false,
-      name: 'Navigator',
-      events: {
-        legendItemClick() {
-          return false;
-        }
-      },
-      unitsLabelShort: null,
-      seasonallyAdjusted: null,
-      pseudoZones: null,
-      visible: true
-    });
-    return chartSeries;
   }
 }
